pyproject.toml
[project]
name = "ollama-test"
version = "0.1.0"
description = "Agent implementation with local Ollama Provider"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "loguru>=0.7.0",
    "pydantic==2.12.*",
    "python-dotenv>=1.2.0",
    "mcp_server_git==2025.12.*",
    "mcp[cli]==1.23.*",
    "ollama==0.6.*",
    "langchain==1.2.*",
    "langchain-core==1.2.*",
    "langchain-ollama==1.0.*",
    "langchain-mcp-adapters==0.2.*",
    "pypdfium2==5.3.*",
    "pillow==12.1.*",
]